pyproject.toml
[project]
name = "azure-codex-proxy"
version = "0.1.0"
description = "Proxy service that refreshes Azure tokens for codex-cli"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity",
    "fastapi",
    "httpx",
    "uvicorn",
]

[project.scripts]
codex-azure = "codex_azure.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["codex_azure"]
